@@ -1,11 +1,11 @@
 name: data_tables
 description: ListView on Mobile and Stateless Data Tables on Tablets and Desktops.
-version: 0.2.4
+version: 1.0.1
 author: Rody Davis <rody.davis.jr@gmail.com>
 homepage: https://github.com/AppleEducate/plugins
 
 environment:
-  sdk: ">=2.0.0 <3.0.0"
+  sdk: ">=2.2.2 <3.0.0"
 
 dependencies:
   flutter:
