@@ -1,19 +1,18 @@
 name: dynamic_tabs
 description: A Flutter plugin for letting the user edit the bottom tabs like the Curpertino iTunes App.
-version: 0.2.0
+version: 1.0.0
 author: Rody Davis <rody.davis.jr@gmail.com>
 homepage: https://github.com/AppleEducate/plugins
 
 environment:
-  sdk: ">=2.0.0 <3.0.0"
+  sdk: ">=2.2.2 <3.0.0"
 
 dependencies:
   flutter:
     sdk: flutter
-  shared_preferences: ^0.5.1+1
-  auto_size_text: ^1.0.0
-  cupertino_controllers: ^0.0.3
-  path_provider: ^0.5.0+1
+  localstorage: ^1.3.0
+  auto_size_text: ^1.1.2
+  cupertino_controllers: ^0.1.0
 
 dev_dependencies:
   flutter_test:
