@@ -3,7 +3,33 @@ import 'package:flutter/material.dart';
 
 import 'data/dessert.dart';
 
-void main() => runApp(MyApp());
+import 'dart:io' show Platform;
+import 'package:flutter/foundation.dart';
+
+// The existing imports
+// !! Keep your existing impots here !!
+
+/// main is entry point of Flutter application
+void main() {
+  // Desktop platforms aren't a valid platform.
+  _setTargetPlatformForDesktop();
+  return runApp(MyApp());
+}
+
+/// If the current platform is desktop, override the default platform to
+/// a supported platform (iOS for macOS, Android for Linux and Windows).
+/// Otherwise, do nothing.
+void _setTargetPlatformForDesktop() {
+  TargetPlatform targetPlatform;
+  if (Platform.isMacOS) {
+    targetPlatform = TargetPlatform.iOS;
+  } else if (Platform.isLinux || Platform.isWindows) {
+    targetPlatform = TargetPlatform.android;
+  }
+  if (targetPlatform != null) {
+    debugDefaultTargetPlatformOverride = targetPlatform;
+  }
+}
 
 class MyApp extends StatefulWidget {
   @override
@@ -129,12 +155,12 @@ class _MyAppState extends State<MyApp> {
             });
             print("New Rows: $value");
           },
-          // mobileItemBuilder: (BuildContext context, int index) {
-          //   final i = _desserts[index];
-          //   return ListTile(
-          //     title: Text(i?.name),
-          //   );
-          // },
+          mobileItemBuilder: (BuildContext context, int index) {
+            final i = _desserts[index];
+            return ListTile(
+              title: Text(i?.name),
+            );
+          },
           onSelectAll: (bool value) {
             for (var row in _items) {
               setState(() {
