@@ -1,17 +1,10 @@
 package com.appleeducate.fluttersms;
 
-import android.Manifest;
 import android.app.Activity;
 import android.content.Intent;
+import android.content.pm.ActivityInfo;
 import android.content.pm.PackageManager;
 import android.net.Uri;
-import android.provider.Settings;
-import android.util.Log;
-
-import java.util.ArrayList;
-import java.util.Arrays;
-
-import androidx.core.app.ActivityCompat;
 import io.flutter.plugin.common.MethodCall;
 import io.flutter.plugin.common.MethodChannel;
 import io.flutter.plugin.common.MethodChannel.MethodCallHandler;
@@ -31,10 +24,20 @@ public class FlutterSmsPlugin implements MethodCallHandler {
   @Override
   public void onMethodCall(MethodCall call, Result result) {
     if (call.method.equals("sendSMS")) {
+      if (!canSendSMS()) {
+        result.error(
+            "device_not_capable",
+            "The current device is not capable of sending text messages.",
+            "A device may be unable to send messages if it does not support messaging or if it is not currently configured to send messages. This only applies to the ability to send text messages via iMessage, SMS, and MMS.");
+        return;
+      }
+
       String message = call.argument("message");
       String recipients = call.argument("recipients");
       sendSMS(recipients, message);
-      result.success("SMS Sent!" );
+      result.success("SMS Sent!");
+    } else if (call.method.equals("canSendSMS")) {
+      result.success(canSendSMS());
     } else {
       result.notImplemented();
     }
@@ -44,13 +47,25 @@ public class FlutterSmsPlugin implements MethodCallHandler {
     this.activity = activity;
   }
 
+  private boolean canSendSMS() {
+    if (!activity.getPackageManager().hasSystemFeature(PackageManager.FEATURE_TELEPHONY))
+      return false;
+
+    Intent intent = new Intent(Intent.ACTION_SENDTO);
+    intent.setData(Uri.parse("smsto:"));
+    ActivityInfo activityInfo =
+        intent.resolveActivityInfo(activity.getPackageManager(), intent.getFlags());
+    if (activityInfo == null || !activityInfo.exported) return false;
+
+    return true;
+  }
+
   private void sendSMS(String phones, String message) {
-     Intent intent = new Intent(Intent.ACTION_VIEW);
-     intent.setData(Uri.parse("smsto:" + phones));
-     intent.putExtra("sms_body", message);
-//     intent.putExtra(Intent.EXTRA_STREAM, attachment);
-     if (intent.resolveActivity( activity.getPackageManager()) != null) {
-       activity.startActivity(intent);
-     }
+    Intent intent = new Intent(Intent.ACTION_SENDTO);
+    intent.setData(Uri.parse("smsto:" + phones));
+    intent.putExtra("sms_body", message);
+    intent.putExtra(Intent.EXTRA_TEXT, message);
+    //     intent.putExtra(Intent.EXTRA_STREAM, attachment);
+    activity.startActivity(intent);
   }
 }
