@@ -3,7 +3,34 @@ import 'package:flutter/material.dart';
 import 'package:persist_theme/persist_theme.dart';
 import 'package:scoped_model/scoped_model.dart';
 
-void main() => runApp(MyApp());
+import 'dart:io' show Platform;
+import 'package:flutter/foundation.dart';
+
+// The existing imports
+// !! Keep your existing impots here !!
+
+/// main is entry point of Flutter application
+void main() {
+  // Desktop platforms aren't a valid platform.
+  _setTargetPlatformForDesktop();
+
+  return runApp(MyApp());
+}
+
+/// If the current platform is desktop, override the default platform to
+/// a supported platform (iOS for macOS, Android for Linux and Windows).
+/// Otherwise, do nothing.
+void _setTargetPlatformForDesktop() {
+  TargetPlatform targetPlatform;
+  if (Platform.isMacOS) {
+    targetPlatform = TargetPlatform.iOS;
+  } else if (Platform.isLinux || Platform.isWindows) {
+    targetPlatform = TargetPlatform.android;
+  }
+  if (targetPlatform != null) {
+    debugDefaultTargetPlatformOverride = targetPlatform;
+  }
+}
 
 class MyApp extends StatefulWidget {
   @override
@@ -16,7 +43,7 @@ class _MyAppState extends State<MyApp> {
   @override
   void initState() {
     try {
-      _model.loadFromDisk();
+      _model.init();
     } catch (e) {
       print("Error Loading Theme: $e");
     }
@@ -30,6 +57,7 @@ class _MyAppState extends State<MyApp> {
       child: new ScopedModelDescendant<ThemeModel>(
         builder: (context, child, model) => MaterialApp(
               theme: model.theme,
+              darkTheme: model.darkTheme,
               home: HomeScreen(),
             ),
       ),
@@ -40,6 +68,7 @@ class _MyAppState extends State<MyApp> {
 class HomeScreen extends StatelessWidget {
   @override
   Widget build(BuildContext context) {
+    final _theme = ScopedModel.of<ThemeModel>(context, rebuildOnChange: true);
     return Scaffold(
       appBar: AppBar(
         title: const Text('Persist Theme'),
@@ -74,6 +103,7 @@ class HomeScreen extends StatelessWidget {
               ],
       ),
       floatingActionButton: FloatingActionButton(
+        backgroundColor: _theme.accentColor,
         child: Icon(Icons.add),
         onPressed: () {},
       ),
