@@ -2,7 +2,34 @@ import 'package:flutter/material.dart';
 
 import 'package:responsive_scaffold/responsive_scaffold.dart';
 
-void main() => runApp(MyApp());
+import 'dart:io' show Platform;
+import 'package:flutter/foundation.dart';
+
+// The existing imports
+// !! Keep your existing impots here !!
+
+/// main is entry point of Flutter application
+void main() {
+  // Desktop platforms aren't a valid platform.
+  _setTargetPlatformForDesktop();
+
+  return runApp(MyApp());
+}
+
+/// If the current platform is desktop, override the default platform to
+/// a supported platform (iOS for macOS, Android for Linux and Windows).
+/// Otherwise, do nothing.
+void _setTargetPlatformForDesktop() {
+  TargetPlatform targetPlatform;
+  if (Platform.isMacOS) {
+    targetPlatform = TargetPlatform.iOS;
+  } else if (Platform.isLinux || Platform.isWindows) {
+    targetPlatform = TargetPlatform.android;
+  }
+  if (targetPlatform != null) {
+    debugDefaultTargetPlatformOverride = targetPlatform;
+  }
+}
 
 class MyApp extends StatefulWidget {
   @override
@@ -12,78 +39,52 @@ class MyApp extends StatefulWidget {
 class _MyAppState extends State<MyApp> {
   var _scaffoldKey = new GlobalKey<ScaffoldState>();
 
+  List<String> _items;
+
+  @override
+  void initState() {
+    _items = List.generate(20, (int index) => "test_$index");
+    super.initState();
+  }
+
   @override
   Widget build(BuildContext context) {
     return MaterialApp(
       home: ResponsiveScaffold.builder(
         scaffoldKey: _scaffoldKey,
         detailBuilder: (BuildContext context, int index, bool tablet) {
+          final i = _items[index];
           return DetailsScreen(
-            // appBar: AppBar(
-            //   elevation: 0.0,
-            //   title: Text("Details"),
-            //   actions: [
-            //     IconButton(
-            //       icon: Icon(Icons.share),
-            //       onPressed: () {},
-            //     ),
-            //     IconButton(
-            //       icon: Icon(Icons.delete),
-            //       onPressed: () {
-            //         if (!tablet) Navigator.of(context).pop();
-            //       },
-            //     ),
-            //   ],
-            // ),
-            body: Scaffold(
-              appBar: AppBar(
-                elevation: 0.0,
-                title: Text("Details"),
-                actions: [
-                  IconButton(
-                    icon: Icon(Icons.share),
-                    onPressed: () {},
-                  ),
-                  IconButton(
-                    icon: Icon(Icons.delete),
-                    onPressed: () {
-                      if (!tablet) Navigator.of(context).pop();
-                    },
-                  ),
-                ],
-              ),
-              bottomNavigationBar: BottomAppBar(
-                elevation: 0.0,
-                child: Container(
-                  child: IconButton(
-                    icon: Icon(Icons.share),
-                    onPressed: () {},
-                  ),
-                ),
-              ),
-              body: Container(
-                child: Center(
-                  child: Text("Item: $index"),
-                ),
-              ),
+            body: new ExampleDetailsScreen(
+              items: _items,
+              row: i,
+              tablet: tablet,
+              onDelete: () {
+                setState(() {
+                  _items.removeAt(index);
+                });
+                if (!tablet) Navigator.of(context).pop();
+              },
+              onChanged: (String value) {
+                setState(() {
+                  _items[index] = value;
+                });
+              },
             ),
           );
         },
+        nullItems: Center(child: CircularProgressIndicator()),
+        emptyItems: Center(child: Text("No Items Found")),
         slivers: <Widget>[
           SliverAppBar(
             title: Text("App Bar"),
           ),
         ],
-        // bottomNavigationBar: BottomAppBar(
-        //   child: Container(
-        //     height: 50.0,
-        //     color: Colors.green,
-        //   ),
-        // ),
-        itemCount: 100,
+        itemCount: _items?.length ?? 0,
         itemBuilder: (BuildContext context, int index) {
+          final i = _items[index];
           return ListTile(
-            leading: Text(index.toString()),
+            leading: Text(i),
           );
         },
         bottomNavigationBar: BottomAppBar(
@@ -102,6 +103,61 @@ class _MyAppState extends State<MyApp> {
               content: Text("Snackbar!"),
             ));
           },
+        ),
+      ),
+    );
+  }
+}
+
+class ExampleDetailsScreen extends StatelessWidget {
+  const ExampleDetailsScreen({
+    Key key,
+    @required List<String> items,
+    @required this.row,
+    @required this.tablet,
+    @required this.onDelete,
+    @required this.onChanged,
+  })  : _items = items,
+        super(key: key);
+
+  final List<String> _items;
+  final String row;
+  final bool tablet;
+  final VoidCallback onDelete;
+  final ValueChanged<String> onChanged;
+
+  @override
+  Widget build(BuildContext context) {
+    return Scaffold(
+      appBar: AppBar(
+        elevation: 0.0,
+        automaticallyImplyLeading: !tablet,
+        title: Text("Details"),
+        actions: [
+          IconButton(
+            icon: Icon(Icons.share),
+            onPressed: () {
+              onChanged(row + " " + DateTime.now().toString());
+            },
+          ),
+          IconButton(
+            icon: Icon(Icons.delete),
+            onPressed: onDelete,
+          ),
+        ],
+      ),
+      bottomNavigationBar: BottomAppBar(
+        elevation: 0.0,
+        child: Container(
+          child: IconButton(
+            icon: Icon(Icons.share),
+            onPressed: () {},
+          ),
+        ),
+      ),
+      body: Container(
+        child: Center(
+          child: Text("Item: $row"),
         ),
       ),
     );
